requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
pyjwt==2.9.0
bcrypt==4.2.0
flask-cors==4.0.1 